perkuat autorisasi website dengan menambah pemberian role di token serta pengecekan role di controller dan route

+++ server/controllers/user.controller.js
@@ -37,6 +37,11 @@ export const getUser = async (req, res, next) => {
 
 export const updateUser = async (req, res, next) => {
     try {
+        //cek apakah admin atau user sendiri
+        if (req.user.role !== 'admin' && req.user.id !== req.params.id) {
+            return next({ status: 403, message: 'Forbidden: Anda tidak punya izin mengubah data user ini' });
+        }
+
         if (req.body.password) {
             req.body.password = await bcrypt.hash(req.body.password, 10);
         }
@@ -60,6 +65,10 @@ export const updateUser = async (req, res, next) => {
 
 export const deleteUser = async (req, res, next) => {
     try {
+        if (req.user.role !== 'admin' && req.user.id !== req.params.id) {
+            return next({ status: 403, message: 'Forbidden: Anda tidak punya izin menghapus user ini' });
+        }
+
         const query = { _id: new ObjectId(req.params.id) };
         await collection.deleteOne(query);
         res.status(200).json({ message: "User has been deleted!" });
